test: Move DummyCompileTime and DummyRuntime to testCore
