+++ main/test/src/io/github/iltotore/iron/test/main/package.scala
@@ -4,18 +4,6 @@ import io.github.iltotore.iron.constraint.Constraint
 
 package object main {
 
-  trait DummyCompileTime
-
-  inline given Constraint.CompileTimeOnly[Boolean, DummyCompileTime] with {
-    override inline def assert(value: Boolean): Boolean = value
-  }
-
-  trait DummyRuntime
-
-  inline given Constraint.RuntimeOnly[Boolean, DummyRuntime] with {
-    override inline def assert(value: Boolean): Boolean = value
-  }
-
   trait Positive
 
   inline given Constraint[Int, Positive] with {
