+++ testCore/src/io/github/iltotore/iron/test/package.scala
@@ -9,4 +9,16 @@ package object test {
   inline given Constraint[Boolean, Dummy] with {
     override inline def assert(value: Boolean): Boolean = value
   }
+
+  trait DummyCompileTime
+
+  inline given Constraint.CompileTimeOnly[Boolean, DummyCompileTime] with {
+    override inline def assert(value: Boolean): Boolean = value
+  }
+
+  trait DummyRuntime
+
+  inline given Constraint.RuntimeOnly[Boolean, DummyRuntime] with {
+    override inline def assert(value: Boolean): Boolean = value
+  }
 }
